\name{ursa-package}

\alias{ursa-package}
\alias{ursula}
\alias{about}
\alias{index}

\docType{package}

\title{ 
Overview
}

\description{
Have a great work with \pkg{ursa}!
}

\details{
See table of content.
}

\author{
Nikita Platonov \email{platonov@sev-in.ru}
}

%%~ \section{Acknowledgements}{
%%~ Thank you.
%%~ }
\keyword{package}
\keyword{spatial}
