\name{ursa-package}

\alias{ursa-package}

\docType{package}

\title{ 
Overview
}

\description{
Some description.
}

\details{
 Details
}

\author{
Nikita Platonov \email{platonov@sevin.ru}
}

%%~ \section{Acknowledgements}{
%%~ Thank you.
%%~ }
\keyword{package}
\keyword{spatial}
