% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Megrez.R
\name{scCNVPip}
\alias{scCNVPip}
\title{Run a post-quantification 10X technology-format single-cell CNV (scCNV) pipeline}
\usage{
scCNVPip(
  project_name = "Ursa_scCNV",
  input_dir = "./",
  output_dir = "./",
  pheno_file,
  confidence_threshold = 50,
  size_threshold = 2e+06,
  cnv_cell_threshold = 5
)
}
\arguments{
\item{project_name}{Project name. 'Ursa_scCNV' by default.}

\item{input_dir}{Directory to all input files. Current working directory by default.}

\item{output_dir}{Output directory. Current working directory by default.
A new folder with the given project name with time stamp as suffix will be
created under the specified output directory.}

\item{pheno_file}{Meta data file directory. Accept only .csv/.txt format files.}

\item{confidence_threshold}{CNV event confidence threshold. Default is 50.}

\item{size_threshold}{CNV event length. Default to 2Mbp in size.}

\item{cnv_cell_threshold}{Filtering out CNV events that are present in less than N% of all cells.
Values range from 0 to 100. Default to 5.}
}
\description{
This function will run a bioinformatics analysis of post-quantification scCNV
pipeline. Supports multiple samples analysis.
}
