% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dubhe.R
\name{scImmunePip}
\alias{scImmunePip}
\title{Run a post-quantification 10X technology-format scImmune Profiling pipeline}
\usage{
scImmunePip(
  project_name = "Ursa_scImmune",
  input_dir = "./",
  output_dir = "./",
  pheno_file,
  integration_method = c("harmony", "seurat"),
  auto_move = T
)
}
\arguments{
\item{project_name}{Project name. 'Ursa_scImmune' by default.}

\item{input_dir}{Directory to all input files. Current working directory by
default.}

\item{output_dir}{Output directory. Current working directory by default.
A new folder with the given project name with time stamp as suffix will be
created under the specified output directory.}

\item{pheno_file}{Meta data file directory. Accept only .csv/.txt format
files.}

\item{integration_method}{Integration method for combining scRNASeq data.
Default by harmony.}
}
\description{
This function will run a bioinformatics analysis of post-quantification scImmune
Profiling pipeline. Supports multiple samples analysis.
}
