% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alioth.R
\name{scRNASEQPip}
\alias{scRNASEQPip}
\title{Run a post-quantification 10X technology-format scRNASEQ pipeline}
\usage{
scRNASEQPip(
  project_name = "Ursa_scRNASEQ",
  input_dir = "./",
  output_dir = "./",
  pheno_file,
  num_genes_lower_bound = 200,
  num_genes_upper_bound = 25000,
  mito_cutoff = 5,
  integration_method = "Seurat",
  cc_regression = 0
)
}
\arguments{
\item{project_name}{Project name. 'Ursa_scRNASEQ' by default.}

\item{input_dir}{Directory to all input files. Current working directory by
default.}

\item{output_dir}{Output directory. Current working directory by default.
A new folder with the given project name with time stamp as suffix will be
created under the specified output directory.}

\item{pheno_file}{Meta data file directory. Accept only .csv/.txt format
files.}

\item{num_genes_lower_bound}{Lower bound to select for cells with number of genes more than this value. Default is 200.}

\item{num_genes_upper_bound}{Upper bound to select for cells with number of genes less than or equals to this value Default is 25000.}

\item{mito_cutoff}{Threshold percentage to select for cells with percentage of mitochondria genes less than this value. This is remove cells which are debris or artefacts and doublets. Default is 5.}

\item{integration_method}{Integration method. Accepts 'Seurat' or "Harmony'
integration. Default is Seurat.}

\item{cc_regression}{Cell cycle regression method. Accepts 0 for no regression,
1 for regression with two-phases, 2 for regression with phase difference (See Seurat).
Default is 0.}
}
\description{
This function will run a bioinformatics analysis of post-quantification scRNASEQ
pipeline. Supports multiple samples analysis.
}
